+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/handler/resetter/NeverResetter.java
@@ -1,0 +1,26 @@
+package com.dwarfeng.voucher.impl.handler.resetter;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * 永远不执行重置的重置器。
+ *
+ * @author DwArFeng
+ * @since beta-1.0.0
+ */
+@Component
+public class NeverResetter extends AbstractResetter {
+
+    @Override
+    protected void doStart() {
+    }
+
+    @Override
+    protected void doStop() {
+    }
+
+    @Override
+    public String toString() {
+        return "NeverResetter{}";
+    }
+}